Implement end-to-end story for development and webJobs nodes
* Hide Files and LogFiles node
* Refactor nodes into own folder

* Extend functionality of WebJob and DeploymentSlots

Adds the functionality to refresh and open in portal to both WebJobs and
DeploymentSlots Nodes.  If not using a Standard subscription, portal
gets hung up as you cannot create webJobs in a Basic sub

* NodeBase owns OpenInPortal function

To prevent the command pallete to have multiple OpenInPortal functions,
the NodeBase now has an optional function and then it's overridden
method is invoked

* Merge rebase

Rebase telemetry merge

* Wizard InputBox improvements #48 & #49  (#50)

Fix:
#48 Wizard input box validation throw error when value is null
#49 Add ignoreFocusOut to wizard input box

* Hide Files and LogFiles nodes

* Refactor nodes into own folder

* Extend functionality of WebJob and DeploymentSlots

Adds the functionality to refresh and open in portal to both WebJobs and
DeploymentSlots Nodes.  If not using a Standard subscription, portal
gets hung up as you cannot create webJobs in a Basic sub

* NodeBase owns OpenInPortal function

To prevent the command pallete to have multiple OpenInPortal functions,
the NodeBase now has an optional function and then it's overridden
method is invoked

* Merge rebase
Rebase telemetry merge